+++ stroom-core-client/src/main/java/stroom/entity/client/presenter/DocumentEditPresenter.java
@@ -71,7 +71,8 @@ public abstract class DocumentEditPresenter<V extends View, D> extends MyPresent
         // Check document permissions if we haven't already.
         if (!checkedPermissions) {
             checkedPermissions = true;
-            securityContext.hasDocumentPermission(docRef.getType(), docRef.getUuid(), DocumentPermissionNames.UPDATE).onSuccess(this::setAllowUpdate);
+            securityContext.hasDocumentPermission(docRef.getType(), docRef.getUuid(), DocumentPermissionNames.UPDATE)
+                    .onSuccess(this::setAllowUpdate);
         }
 
         this.entity = entity;