+++ stroom-docstore/stroom-docstore-impl/src/main/java/stroom/docstore/impl/StoreImpl.java
@@ -375,6 +375,7 @@ public class StoreImpl<D extends Doc> implements Store<D> {
                 }
             });
         } catch (final IOException e) {
+            LOGGER.error(LogUtil.message("Error serialising {}", document.getType()), e);
             throw new UncheckedIOException(e);
         }